requirements.txt
# Core
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1
psycopg2-binary==2.9.9

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Redis
redis==5.0.1
aioredis==2.0.1

# Environment
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Email
email-validator==2.1.0

# CORS
starlette==0.27.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0